Dev_2/requirements.txt
fastapi
uvicorn
httpx
feedparser
scikit-learn
python-dateutil
pydantic
